Added Prometheus and Grafana to docker-compose

+++ src/main/com/stockcomp/response/RealTimePriceResponse.kt
@@ -5,25 +5,25 @@ import com.fasterxml.jackson.annotation.JsonProperty
 class RealTimePriceResponse {
     @get:JsonProperty("currentPrice")
     @set:JsonProperty("c")
-    var currentPrice: Int? = null
+    var currentPrice: Double? = null
 
     @get:JsonProperty("highPrice")
     @set:JsonProperty("h")
-    var highPrice: Int? = null
+    var highPrice: Double? = null
 
     @get:JsonProperty("lowPrice")
     @set:JsonProperty("l")
-    var lowPrice: Int? = null
+    var lowPrice: Double? = null
 
     @get:JsonProperty("openPrice")
     @set:JsonProperty("o")
-    var openPrice: Int? = null
+    var openPrice: Double? = null
 
     @get:JsonProperty("previousClosePrice")
     @set:JsonProperty("pc")
-    var previousClosePrice: Int? = null
+    var previousClosePrice: Double? = null
 
     @get:JsonProperty("time")
     @set:JsonProperty("t")
-    var time: Int? = null
+    var time: Long? = null
 }
